Overhaul build matrix generation
The previous approach to generating the build matrix was based on
the GitHub compare API. However, that API has the limitation of only
providing a maximum of 300 changed files and otherwise truncates the
result.

A single pull request may change more than 300 files, especially when
removing old images. This leads to erroneous generation of the build
matrix when images that have been deleted in such a pull request are
selected to be built.

Use a different technique to build the matrix: Use GitHub's Git tree API
to recursively retrieve all tree hashes for the base and target commits.
Then, select all the tree hashes pointing to an image folder and build
all the images from the target commit that lack corresponding hashes in
the base commit.

The Git tree API is limited to 100,000 entries. This gives enough
headroom for the future :-)

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>